TAC-1578 TAC-1585 View Rating In Tenant Profile

See merge request tachyonhub/tachyon-core!533

+++ aspnet-core/src/TACHYON.Application.Shared/Authorization/Users/Profile/Dto/TenantProfileInformationDto.cs
@@ -20,6 +20,6 @@ namespace TACHYON.Authorization.Users.Profile.Dto
 
         public int CityId { get; set; }
 
-        public double Rating { get; set; }
+        public decimal Rating { get; set; }
     }
 }
